requirements.txt
# accelerate==1.5.2
# annotated-types==0.7.0
# anthropic==0.49.0
# anyio==4.7.0
# certifi==2024.12.14
# charset-normalizer==3.4.0
# contourpy==1.3.2
# cycler==0.12.1
# defusedxml==0.7.1
# distro==1.9.0
# exceptiongroup==1.2.2
# filelock==3.18.0
# fonttools==4.57.0
# fsspec==2025.3.0
# h11==0.14.0
# httpcore==1.0.7
# httpx==0.28.1
# huggingface-hub==0.29.3
# idna==3.10
# iniconfig==2.1.0
# Jinja2==3.1.6
# jiter==0.8.2
# joblib==1.4.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# kiwisolver==1.4.8
langchain-anthropic==0.2.0
langchain-core==0.3.27
langchain-huggingface==0.1.2
langchain-openai==0.2.13
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.4
# nvidia-cublas-cu12==12.4.5.8
# nvidia-cuda-cupti-cu12==12.4.127
# nvidia-cuda-nvrtc-cu12==12.4.127
# nvidia-cuda-runtime-cu12==12.4.127
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.2.1.3
# nvidia-curand-cu12==10.3.5.147
# nvidia-cusolver-cu12==11.6.1.9
# nvidia-cusparse-cu12==12.3.1.170
# nvidia-cusparselt-cu12==0.6.2
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.4.127
# nvidia-nvtx-cu12==12.4.127
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.1
tomli==2.2.1
# torch==2.6.0
tqdm==4.67.1
# transformers==4.50.0
# triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
yapf==0.43.0

